+++ packages/charts/src/utils/themes/amsterdam_dark_theme.ts
@@ -224,6 +224,7 @@ export const AMSTERDAM_DARK_THEME: Theme = {
     margin: 0,
     labelOptions: {
       maxLines: 1,
+      widthLimit: 250,
     },
   },
   crosshair: {
@@ -434,7 +435,7 @@ export const AMSTERDAM_DARK_THEME: Theme = {
     emptyBackground: Colors.Transparent.keyword,
     nonFiniteText: 'N/A',
     minHeight: 64,
-    titleWeight: 'bold',
+    titleWeight: 500,
   },
   bulletGraph: {
     textColor: '#E0E5EE',