+++ packages/charts/src/utils/themes/legacy_dark_theme.ts
@@ -229,6 +229,7 @@ export const LEGACY_DARK_THEME: Theme = {
     margin: 0,
     labelOptions: {
       maxLines: 1,
+      widthLimit: 250,
     },
   },
   crosshair: {
@@ -438,7 +439,7 @@ export const LEGACY_DARK_THEME: Theme = {
     emptyBackground: Colors.Transparent.keyword,
     nonFiniteText: 'N/A',
     minHeight: 64,
-    titleWeight: 'bold',
+    titleWeight: 500,
   },
   bulletGraph: DARK_THEME_BULLET_STYLE,
   tooltip: {