+++ packages/charts/src/utils/themes/light_theme.ts
@@ -247,6 +247,7 @@ export const LIGHT_THEME: Theme = {
     margin: 0,
     labelOptions: {
       maxLines: 1,
+      widthLimit: 250,
     },
   },
   crosshair: {
@@ -456,7 +457,7 @@ export const LIGHT_THEME: Theme = {
     emptyBackground: Colors.Transparent.keyword,
     nonFiniteText: 'N/A',
     minHeight: 64,
-    titleWeight: 'bold',
+    titleWeight: 500,
   },
   bulletGraph: LIGHT_THEME_BULLET_STYLE,
   tooltip: {