+++ storybook/stories/legend/17_tabular_data.story.tsx
@@ -25,6 +25,7 @@ const getLabelOptionKnobs = (): LegendLabelOptions => {
 
   return {
     maxLines: number('max label lines', 1, { min: 0, step: 1 }, group),
+    widthLimit: 250,
   };
 };
 
@@ -52,6 +53,7 @@ const datasets: Record<'defaultDataset' | 'shortCopyDataset' | 'longCopyDataset'
 
 export const Example: ChartsStory = (_, { title, description }) => {
   const hideActions = boolean('Hide legend action', false, 'Legend');
+  const legendActionOnHover = boolean('Show legend action on hover', true, 'Legend');
   const showLegendExtra = !boolean('Hide legend extra', false, 'Legend');
   const showColorPicker = !boolean('Hide color picker', true, 'Legend');
   const legendPosition = customKnobs.enum.position('Legend position', undefined, { group: 'Legend' });
@@ -90,6 +92,7 @@ export const Example: ChartsStory = (_, { title, description }) => {
         baseTheme={useBaseTheme()}
         legendPosition={legendPosition}
         legendAction={hideActions ? undefined : getLegendAction(euiPopoverPosition)}
+        legendActionOnHover={legendActionOnHover}
         legendColorPicker={showColorPicker ? getColorPicker(euiPopoverPosition) : undefined}
         legendSize={getLegendSizeKnob()}
         legendValues={showLegendExtra ? legendValues : []}